// src/sites/item/onenote/operations/item/onenoteOperationItemRequestBuilderDeleteRequestConfiguration.ts
import { type RequestOption } from '@microsoft/kiota-abstractions';

export interface OnenoteOperationItemRequestBuilderDeleteRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}